Finshed campaign class

+++ CampaignFolder/Campaign.cs
@@ -9,11 +9,11 @@ namespace CashRegister.CampaignFolder
     public class Campaign
     {
         public int ProdPLU { get; set; }
-        public int Discount { get; set; }
+        public decimal Discount { get; set; }
         public DateOnly StartDate { get; set; }
         public DateOnly EndDate { get; set; }
 
-        public Campaign(int _prodPLU, int _discount, DateOnly _startDate, DateOnly _endDate)
+        public Campaign(int _prodPLU, decimal _discount, DateOnly _startDate, DateOnly _endDate)
         {
             ProdPLU = _prodPLU;
             Discount = _discount;
